:star2: Expanding material database
Merging in a number of materials from the database in Telewavesim. Added
each specific reference in full, and double-checked the coefficients.
Found two minor differences:

1. Lawsonite - correct moduli, but incorrect matrix position assignment.
2. Harzburgite - very minor difference between previous C15 modulus and
that in literature.

A small number of root references could not be found, namely those that
were originally published in Russian (in the proceedings of the Russian
Academy of Sciences).

It will be worth adding an attribute field to the Material class that
will carry an optional material ID and reference, where appropriate.
This should be put into the __str__ method, so it can be printed at
appropriate stages - always good to remind people that these numbers
have a reference!