:electron: Tests for material database
Add a test for the `anisotropy.materials.database` module. This test
involves reading in an example material (in this case, olivine) then
testing the following are true:

- The material is of type `anisotropy.materials.Material`
- The elastic constants (in the form of the Voigt stiffness tensor) and
  the density are correct
- Finally, ensure that attempting to load an invalid material raises the
  expected Exception (`anisotropy.utils.errors.InvalidMaterialID`)

Additionally, made a minor change to the bond rotation matrix to make it
a bit more readable (could still do with a bit of work).