:wavy_dash: Add C splitting routines
Add implementation of the Silver and Chan, 1991, eigenvalue minimisation
splitting routines. These have been written in C and parallelised over
the window search portion (motivated by a desire to replicate the
approach presented in the SPLIT codes of Teanby).

The implementation is original, with influence/inspiration from the C
routines in multisplit (F. Tilmann). It leverages the GNU scientific
library (GSL) to manipulate the vectors. The promise here is that it is
far easier to manipulate the raw waveform data in Python (pre-processing
steps such as response correction, detrending, application of filters,
window identification, etc.) AND that by parallelising the code, we can
take full advantage of the compute power available to the user.