:floppy_disk: Add GH workflow
Add a GH workflow to automatically build wheels and upload to PyPI. For
now, it only uploads to the Test PyPI. In future, can hopefully expand
this to include proper upload to PyPI. The secret tokens have been set
up in the AnisotroPy repo. Base workflow .yml file was built from a
tutorial, with (as ever) some inspiration from the ObsPy example!