:arrow_up: Add tensor decomposition
Add elastic tensor decomposition after Browaeys and Chevrot (2004).
Currently not 100% true to what is outlined in the paper - it does not
first try to rotate the 6x6 elastic matrix onto a basis that minimises
the number of unique elements (or something along those lines).

Does reproduce the values outlined in the paper (will add test to
confirm this soon), as well as highlighting that some of the values in
the paper are actually wrong. Specifically:

1. element (1, 2) of the decomposed hexagonal component of the enstatite
elastic matrix is 1.7, when it should be 0 (appears to have been carried
over from the same element of the olivine decomposition that precedes it).
2. The orthorhombic component of the enstatite elastic matrix has been
multiplied by a factor of ten!

Other than that, bonza.