:bookmark_tabs: Add materials module tutorials
Add tutorials for the materials module that cover some of the basics of
loading Materials from literature, building a new Material, mixing
Materials, and evaluating some of the elastic properties of a given
Material. This module is an important basis for the effective medium
modelling code that forms the effective_modelling module.