:wrench: Minor adjustments
A few minor adjustments to fix failing tests, including adding a
warning when the volume fractions are re-normalised.

Another minor adjustment was made to the default parameters of one the
new plotting function, which previously was using [] (not recommended).
These have been swapped for None's and the corresponding test within the
function has been adjusted accordingly.